fix: images location import errors caused by assets moving into images
